Small refactoring before adding new tests
